requirements.txt
attrs==21.2.0
boto3==1.18.39
botocore==1.21.39
jmespath==0.10.0
protobuf==3.15.6
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
